ISP20200319 Local Order(Orders.)則維持原本撈取規則
其它變更準備資料方式,以order_Eachcons準備部位,避免Bundle沒有建立無法檢查到

原本使用store Function：QtyBySetPerSubprocess
改用C# QtyBySetPerSubprocess

調整C# QtyBySetPerSubprocess

變更程式
PPIC P01_ProductionOutput_LoadingoutputDetail
PPIC R13
Subcon P42_SubProcessStatus
Subcon R44
Planning R15

Planning R15 WIP CutQty右邊欄位加入 RFID SewingLine In Qty顯示SewingLine subprcess RFID成衣件數資料。
Status判斷也加上判斷SewingLine

+++ Sci.Production.PPIC/P01_ProductionOutput.cs
@@ -326,7 +326,7 @@ where oq.id = '{this.masterData["ID"]}'
                 return;
             }
 
-            this.numLoadingQty.Value = MyUtility.Convert.GetInt(loadingoutput.Compute("sum(AccuInCome)", string.Empty));  // Sewing Q'ty
+            this.numLoadingQty.Value = MyUtility.Convert.GetDecimal(loadingoutput.Compute("sum(AccuInCome)", string.Empty));  // Sewing Q'ty
 
             this.listControlBindingSource1.DataSource = sewingData;
             this.listControlBindingSource2.DataSource = cuttingData;
